QCMAP: Packet counters are not updated in conntrack
Return value of counter api returns s64 instead of u64
Type-casting to the right type.

Acked-by: Vinisha Varre <vvarre@qti.qualcomm.com>

+++ net/netfilter/nf_conntrack_netlink.c
@@ -245,8 +245,8 @@ dump_counters(struct sk_buff *skb, struct nf_conn_acct *acct,
 		pkts = atomic64_xchg(&counter[dir].packets, 0);
 		bytes = atomic64_xchg(&counter[dir].bytes, 0);
 	} else {
-		pkts = atomic64_read(&counter[dir].packets);
-		bytes = atomic64_read(&counter[dir].bytes);
+		pkts = (u64)atomic64_read(&counter[dir].packets);
+		bytes = (u64)atomic64_read(&counter[dir].bytes);
 	}
 
 	nest_count = nla_nest_start(skb, attr);
